requirements.txt
langchain==1.1.3
langchain-core==1.1.3
langchain-groq==1.1.0

sentence-transformers==5.1.2

PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
transformers==4.57.3
torch==2.9.1
fastapi==0.124.4
uvicorn[standard]
python-multipart==0.0.21